+++ DynamicSun.Domain/Abstractions/IWeather.cs
@@ -14,7 +14,7 @@ namespace DynamicSun.Domain.Abstractions
         public decimal? Temperature { get; set; } // Температура
         public decimal? AirHumidity { get; set; } // Отн. влажность воздуха, %
         public decimal? DewPoint { get; set; } // Точка росы, гр
-        public decimal? Pressure { get; set; } // Атм. давление, мм рт.ст.
+        public int? Pressure { get; set; } // Атм. давление, мм рт.ст.
         public string? Wind { get; set; } // Направление ветра
         public int? WindSpeed { get; set; } // Направление ветра, м/с
         public int? Cloudiness { get; set; } // Обачность, %
